+++ src/screens/home/components/blocks/components/mobile/index.tsx
@@ -37,9 +37,9 @@ const Mobile: React.FC<{
               time={dayjs.utc(x.timestamp).fromNow()}
               leader={(
                 <AvatarName
-                  address={x.leader.address}
-                  imageUrl={x.leader.imageUrl}
-                  name={x.leader.name}
+                  address={x.proposer.address}
+                  imageUrl={x.proposer.imageUrl}
+                  name={x.proposer.name}
                 />
               )}
               hash={getMiddleEllipsis(x.hash, {